+++ wombat/core/data_classes.py
@@ -14,7 +14,6 @@ import attrs
 import numpy as np
 import pandas as pd
 from attrs import Factory, Attribute, field, define
-from scipy.stats import weibull_min
 
 from wombat.utilities.time import HOURS_IN_DAY, HOURS_IN_YEAR, parse_date
 
@@ -514,6 +513,13 @@ class Failure(FromDictMixin):
         only a repair is necessary. Defaults to False
     description : str
         A short text description to be used for logging.
+    rng : np.random._generator.Generator
+
+        .. note:: Do not provide this, it comes from
+            :py:class:`wombat.core.environment.WombatEnvironment`
+
+        The shared random generator used for the Weibull sampling. This is fed through
+        the simulation environment to ensure consistent seeding between simulations.
     """
 
     scale: float = field(converter=float)
@@ -530,16 +536,18 @@ class Failure(FromDictMixin):
         ),
     )
     system_value: int | float = field(converter=float)
+    rng: np.random._generator.Generator = field(
+        eq=False,
+        validator=attrs.validators.instance_of(np.random._generator.Generator),
+    )
     replacement: bool = field(default=False, converter=bool)
     description: str = field(default="failure", converter=str)
-    weibull: weibull_min = field(init=False, eq=False)
     request_id: str = field(init=False)
 
     def __attrs_post_init__(self):
         """Create the actual Weibull distribution and converts equipment requirements
         to a list.
         """
-        object.__setattr__(self, "weibull", weibull_min(self.shape, scale=self.scale))
         object.__setattr__(
             self,
             "service_equipment",
@@ -565,7 +573,7 @@ class Failure(FromDictMixin):
         if self.scale == self.shape == 0:
             return None
 
-        return self.weibull.rvs(size=1)[0] * HOURS_IN_YEAR
+        return self.rng.weibull(self.shape, size=1)[0] * self.scale * HOURS_IN_YEAR
 
     def assign_id(self, request_id: str) -> None:
         """Assign a unique identifier to the request.
@@ -600,8 +608,13 @@ class SubassemblyData(FromDictMixin):
 
     name: str = field(converter=str)
     maintenance: list[Maintenance | dict[str, float | str]]
-    failures: dict[int, Failure | dict[str, float | str]]
+    failures: list[Failure | dict[str, float | str]] | dict[
+        int, Failure | dict[str, float | str]
+    ]
     system_value: int | float = field(converter=float)
+    rng: np.random._generator.Generator = field(
+        validator=attrs.validators.instance_of(np.random._generator.Generator)
+    )
 
     def __attrs_post_init__(self):
         """Convert the maintenance and failure data to ``Maintenance`` and ``Failure``
@@ -624,14 +637,17 @@ class SubassemblyData(FromDictMixin):
             if TYPE_CHECKING:
                 assert isinstance(kwargs, dict)
             kwargs.update({"system_value": self.system_value})
-        object.__setattr__(
-            self,
-            "failures",
-            {
-                level: Failure.from_dict(kw) if isinstance(kw, dict) else kw
-                for level, kw in self.failures.items()
-            },
-        )
+
+        failures_list = []
+        rng_dict = {"rng": self.rng}
+        if TYPE_CHECKING:
+            assert isinstance(self.failures, dict)
+        for config in self.failures.values():
+            if TYPE_CHECKING:
+                assert isinstance(config, dict)
+            config.update(rng_dict)
+            failures_list.append(Failure.from_dict(config))
+        object.__setattr__(self, "failures", failures_list)
 
 
 @define(frozen=True, auto_attribs=True)